FIX: cancel_timer seems to get into an infinite loop
Running into an issue with cancelling a timer and it seems like we get into an infinite loop in cancel_timer() (timer_queue.hpp).

Based on this commit (made by @chriskohlhoff), it appears that setting the error code for completion handler should be done after the item has been popped on the op_queue:
https://github.com/boostorg/asio/commit/a51b9b819844a2c6150b36d8f565063c75a429c4

I believe this fix needs to be applied to the cancel_timer function (i.e. lines 190 and 191 should be swapped).
https://github.com/boostorg/asio/blob/8a77375b02f97c6ecbe35bdf08e457df83e89969/include/boost/asio/detail/timer_queue.hpp#L190
I swapped these lines on my development system and it seems to resolve the infinite loop.

URL: https://github.com/chriskohlhoff/asio/issues/1025